+++ mongoengine/__init__.py
@@ -19,7 +19,7 @@ __all__ = (list(document.__all__) + fields.__all__ + connection.__all__ +
            list(queryset.__all__) + signals.__all__ + roles.__all__ +
            list(errors.__all__))
 
-VERSION = (0, 8, 4)
+VERSION = (0, 8, 6)
 
 
 def get_version():