+++ mongoengine/document.py
@@ -66,7 +66,7 @@ class EmbeddedDocument(BaseDocument):
 
     def __eq__(self, other):
         if isinstance(other, self.__class__):
-            return self._data == other._data
+            return self.to_mongo() == other.to_mongo()
         return False
 
     def __ne__(self, other):